Add ship purchasing and naming
A Ship type was added. Players can purchase them through the
Hangar which may eventually serve as a per-station instance, but
for now is basically a global thing.

Ships currently have no affect on the game but should eventually
influence mission progress / success. Interaction with ships and
hangars also needs to be fleshed out more.