Add money, char command, make config independent
Characters now have money. Ships can be bought and sold and the balance
is reflected in the character's model.

Deleted the 'progress' command in favor of a 'char' command that will
eventually own all things character-related. Made a better 'character
sheet' type output that shows the character state.

The game config was moved out to be independent from main.py. Though
at the moment this read/write to json design is wide open to race
conditions. Need to convert it to be set up in the database.